chore: move info into client/auth
其为 client 运行时产生的结构, 理应放置其下.
同时, 对标 MiraiGo 的 client/internal/auth, 但因为 main 中需要引用, 故不再隐藏
