+++ client/auth/device.go
@@ -1,4 +1,4 @@
-package info
+package auth
 
 import (
 	"encoding/json"
