+++ client/auth/sig.go
@@ -1,4 +1,4 @@
-package info
+package auth
 
 import (
 	"bytes"
