+++ client/event/self.go
@@ -1,11 +1,11 @@
 package event
 
 import (
-	"github.com/LagrangeDev/LagrangeGo/info"
+	"github.com/LagrangeDev/LagrangeGo/client/auth"
 	"github.com/LagrangeDev/LagrangeGo/packets/pb/message"
 )
 
-func ParseSelfRenameEvent(event *message.SelfRenameMsg, sig *info.SigInfo) *Rename {
+func ParseSelfRenameEvent(event *message.SelfRenameMsg, sig *auth.SigInfo) *Rename {
 	sig.Nickname = event.Body.RenameData.NickName
 	return &Rename{
 		Uin:      event.Body.Uin,
