+++ client/ntlogin.go
@@ -5,15 +5,12 @@ import (
 	"net/url"
 	"strconv"
 
-	"github.com/LagrangeDev/LagrangeGo/packets/wtlogin/loginState"
-
-	"github.com/LagrangeDev/LagrangeGo/packets/pb/login"
-
-	"github.com/LagrangeDev/LagrangeGo/utils/crypto"
-
-	"github.com/LagrangeDev/LagrangeGo/info"
+	"github.com/LagrangeDev/LagrangeGo/client/auth"
 	"github.com/LagrangeDev/LagrangeGo/internal/proto"
+	"github.com/LagrangeDev/LagrangeGo/packets/pb/login"
+	"github.com/LagrangeDev/LagrangeGo/packets/wtlogin/loginState"
 	"github.com/LagrangeDev/LagrangeGo/utils"
+	"github.com/LagrangeDev/LagrangeGo/utils/crypto"
 )
 
 var loginLogger = utils.GetLogger("login")
@@ -26,7 +23,7 @@ func buildNtloginCaptchaSubmit(ticket, randStr, aid string) proto.DynamicMessage
 	}
 }
 
-func buildNtloginRequest(uin uint32, app *info.AppInfo, device *info.DeviceInfo, sig *info.SigInfo, credential []byte) ([]byte, error) {
+func buildNtloginRequest(uin uint32, app *auth.AppInfo, device *auth.DeviceInfo, sig *auth.SigInfo, credential []byte) ([]byte, error) {
 	body := proto.DynamicMessage{
 		1: proto.DynamicMessage{
 			1: proto.DynamicMessage{
@@ -69,7 +66,7 @@ func buildNtloginRequest(uin uint32, app *info.AppInfo, device *info.DeviceInfo,
 	}.Encode(), nil
 }
 
-func ParseNtloginResponse(response []byte, sig *info.SigInfo) (loginState.State, error) {
+func ParseNtloginResponse(response []byte, sig *auth.SigInfo) (loginState.State, error) {
 	var frame login.SsoNTLoginEncryptedData
 	err := proto.Unmarshal(response, &frame)
 	if err != nil {
